requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.0
openpyxl>=3.1.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
